Title: 삼 오 무, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 241024/삼 오 무/three-five-moo.cpp
@@ -13,7 +13,7 @@ int main() {
     int n;
     cin >> n;
     
-    int left = 1, right = INT_MAX;
+    long long left = 1, right = INT_MAX;
     long long ans = INT_MAX;
     while (left <= right)
     {
